Pica/Texture: Move part of ETC1 decoding to new file and cleanups
